Move vue back to globally available
