+++ assets/js/components/Example.vue
