+++ views/base.njk
@@ -48,5 +48,8 @@
                 <script type="text/javascript" src={{val}}></script>
             {% endfor %}
         {% endif %}
+        {% if appJs %}
+            <script type="text/javascript" src={{appJs}}></script>
+        {% endif %}
     {% endblock %}
 </html>
