+++ webpack.config.js
@@ -7,6 +7,7 @@ module.exports = (env, argv) => {
     mode: argv.mode,
     entry: {
       styles: './assets/scss/app.scss',
+      app: './assets/js/app.js',
       personal: './routes/personal/js/personal.js',
     },
     output: {
